Initial Addition of Store info
Initially addition of Store info.
Includes Page, and Basic function in the Create bug page.

Things that need to be worked on, is the helping for getting version for each device. Windows is done for Win10 + Win8 (with a method that works on Win7)
All other devives have the pop-up open but do nothing.